requirements.txt
asgiref==3.11.1
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
dj-database-url==3.1.2
Django==6.0.3
greenlet==3.3.2
idna==3.11
mypy_extensions==1.1.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
platformdirs==4.9.4
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
